api/requirements.txt
fastapi
uvicorn[standard]
supabase
pytest
pytest-asyncio
pydantic
python-dotenv
httpx 